feat(cli): move py and ts under generate command
